Fix UML composite and shared association
Using the UML Composite and Share Association
tools was only creating a normal association.
To resolve, copied the SysML implementation,
and refactored some of the functions in to the
UML toolbox module to prevent duplication.
